Refactoring, before I add reduction of high order cells.
We want to have support for the conversion of high
order cells to linear cells. To do this we want
to add a new class that will be used by the data set
converter as input to the MixedCellConnectivity class.

MixedCellConnectivity is already doing to much work, and
it needs to be broken into sub classes that only have a single task rather than this monotholic class that
knows everything.